requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
Django==2.2.1
django-celery==3.2.2
django-cors-headers==3.6.0
django-redis==4.12.1
django-rest-framework==0.1.0
django-tinymce==3.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
importlib-metadata==3.3.0
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
kombu==3.0.37
MarkupSafe==1.1.1
packaging==20.8
Pillow==8.0.1
prompt-toolkit==3.0.8
py3Fdfs==2.2.0
pycparser==2.20
PyMySQL==0.10.1
pyparsing==2.4.7
pytz==2020.4
redis==3.5.3
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
